Navigation Bar
I built the initial layout of the navigation bar for the e-commerce site using HTML and CSS. The logo sits on the left, the main menu is centered, and the search/login icons are on the right. For this commit, I have only needed to update the index file and add some basic CSS. The CSS for all sections from now will be saved in "style.css" which I have chosen to store in its own folder in our project folder named "css".