Product List 1
HTML and CSS are updated to add in the next section "Product List 1". Product images will need to be saved to a folder inside the project folder at some point and renamed This will enable them to be viewed by users, in this section and others like it, where the purpose is to display products being sold on the website.