Sale Banner
HTML and CSS files are updated to include a "Sale Banner". For this to be viewed and used properly, another image needs to be added to the projects "images" folder and used as the background of this section. The dimensions and name used for the background can be found in the banner section in the updated style.css file.