Services and Footer
The last sections of this ecommerce page have been added to the existing HTML and CSS files. This is a "Services" section and then finally, the footer. Customise these sections and previous ones in the finished home page by editing the appropriate details in the HTML file. Changes to the colours that are used, can be made in the CSS file.